# docs/openapi.yaml
openapi: 3.1.0
info:
  description: "This is API overview"
  version: 0.1.0
  title: PyrHouse API
  contact:
    email: warrmag7@gmail.com
  license:
    name: MIT
security:
  - bearerAuth: [] # use the same name as above    
# tags:
#   - name: users
#     description: Operations about user
#     externalDocs:
#       description: Find out more about...
#       url: http://swagger.io
paths:
...
  /auth/discord/callback:
    get:
      tags:
        - auth
      summary: Callback Discord OAuth2
      description: |
        Endpoint wywoływany przez Discord po autoryzacji.
        Wymienia code na token i tworzy/znajduje użytkownika.
        Nowi użytkownicy Discord mają `active=false` i wymagają ręcznej aktywacji przez admina.
      operationId: discordCallback
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: Authorization code z Discord
        - in: query
          name: state
          schema:
            type: string
          required: true
          description: State parameter dla CSRF protection
      responses:
        "200":
          description: Logowanie udane
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token
        "400":
          description: Nieprawidłowy state lub brak code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Konto jest nieaktywne
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Konto jest nieaktywne. Skontaktuj się z administratorem."
        "500":
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /locations:
    post:
      tags:
        - location
      summary: Create new location (Czteropak)
      description: ""
      requestBody:
        content:
          application/json:
            schema:
                type: object
                properties:
                  name:
                    type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "400":
          description: Invalid payload
        "409":
          description: Name already taken
  /locations/{locationID}/assets: 
    get:
      tags:
        - location
      summary: Get list of assets and stock items in given location to-refactor path
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
                  stock_items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Stock"
        "500":
          description: todo error handling
          $ref: "#/components/responses/DefaultError"
  /locations/{locationID}:
    delete:
      tags:
        - location
      summary: Remove location by ID
      parameters:
        - in: path
          name: locationID
          schema:
            type: integer
          required: true
          description: Numeric ID of the location
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: "Location deleted successfully"
        "409":
          description: Conflict - Cannot delete location with existing items
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Could not delete location"
                  details:
                    type: string
        "500":
          $ref: "#/components/responses/DefaultError"
  /locations/{locationID}/search:
    get:
      tags:
        - location
      summary: Search items in location
      description: Search for items in a specific location by serial number, category name, or category label
      parameters:
        - in: path
          name: locationID
          schema:
            type: integer
          required: true
          description: Numeric ID of the location
        - in: query
          name: q
          schema:
            type: string
          required: true
          description: Search query (minimum 1 character)
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "400":
          description: Invalid search query
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Search query must be at least 1 character long"
        "500":
          $ref: "#/components/responses/DefaultError"
...
  /users/{id}/link-discord:
    parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags:
        - users
      summary: Połącz konto z Discord
      description: |
        Łączy istniejące konto użytkownika z kontem Discord.
        Wymaga autoryzacji - tylko właściciel konta lub admin może wykonać tę operację.
        Frontend musi najpierw uzyskać code i state z Discord OAuth flow.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - state
              properties:
                code:
                  type: string
                  description: Authorization code z Discord OAuth
                state:
                  type: string
                  description: State parameter z Discord OAuth
        required: true
      responses:
        "200":
          description: Konto Discord zostało połączone
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Konto Discord zostało połączone"
        "400":
          description: Nieprawidłowe dane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Brak uprawnień
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Konto Discord jest już połączone z innym użytkownikiem
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "To konto Discord jest już połączone z innym użytkownikiem"
        "500":
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: pyrhouse-backend-f26ml.ondigitalocean.app/api
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for 
  responses:
    DefaultError:
      description: Operation failed location has stock/assets already
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Could not delete location
        details:
          type: string
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 321
        serial:
          type: string
          example: XYZ321
        pyrcode:
          type: string
          example: PYR0L100
        location:
          $ref: "#/components/schemas/Location"
        category:
          $ref: "#/components/schemas/ItemCategory"
    Stock:
      type: object
      properties:
        id:
          type: integer
        category:
          $ref: "#/components/schemas/ItemCategory"
        quantity:
          type: integer
    ItemCategory:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          example: asset 
        name:
          type: string
          example: laptop       
        label:
          type: string
          example: Laptop
    Location:
      type: object
      properties:
        id: 
          type: integer
          example: 3
        name:
          type: string
          example: "Gzdaczroom"
...